Address PR #9 review nits
- PlaybackModule: File.deleteOnExit() only removes *empty* directories,
  so the staged GTFS copy leaked across runs. Replaced with a shutdown
  hook that recursively deletes the staging root on JVM exit.
- tools/wmata_capture/capture.py: dropped a redundant int() around
  round() — round(float) already returns int in Python 3.
- tools/wmata_capture/README.md: reworded the confusing "restore @Ignore
  removal into a real code change" bullet; the intent is simply that
  un-ignoring the test is part of the refresh, not a local throwaway.

Skipped the rest: the ehcache pool-disk reference mirrors a pre-existing
pattern in transitclock/ and transitclockPipelineTests/, so changing it
only in the integration file would diverge from the "shadow of core"
design. The run-all-tests `-am` comment is incorrect — the root pom's
top-level <modules> list is always in the reactor, so core builds
alongside the profile's extras (verified earlier). And the README code
fence already matches the file's style (no fences in this README use
language specifiers).

+++ tools/wmata_capture/capture.py
@@ -264,7 +264,7 @@ def record_vehicle(
     lon = getattr(pos, "longitude", None)
     if lat is None or lon is None:
         return
-    heading = int(round(getattr(pos, "bearing", 0.0) or 0.0))
+    heading = round(getattr(pos, "bearing", 0.0) or 0.0)
 
     ts = int(getattr(vehicle_msg, "timestamp", 0) or 0)
     if ts <= 0:
